+++ rust/headless-client/src/interface.rs
@@ -1,0 +1,13 @@
+//! A cross-platform virtual network interface that can control DNS
+
+#[cfg(target_os = "linux")]
+pub mod linux;
+#[cfg(target_os = "linux")]
+pub use linux as platform;
+
+#[cfg(target_os = "windows")]
+pub mod windows;
+#[cfg(target_os = "windows")]
+pub use windows as platform;
+
+pub(crate) use platform::InterfaceManager;
