+++ rust/headless-client/src/lib.rs
@@ -42,6 +42,8 @@ pub mod windows;
 #[cfg(target_os = "windows")]
 pub use windows as platform;
 
+mod interface;
+
 /// Only used on Linux
 pub const FIREZONE_GROUP: &str = "firezone-client";
 
@@ -413,7 +415,7 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
     let (cb_tx, mut cb_rx) = mpsc::channel(100);
 
     let send_task = tokio::spawn(async move {
-        let mut interface = platform::InterfaceManager::new()?;
+        let mut interface = interface::InterfaceManager::new()?;
 
         while let Some(msg) = cb_rx.recv().await {
             match msg {
